todolist/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.0.0
Faker==12.0.1
idna==3.3
importlib-metadata==4.10.1
install==1.3.5
Markdown==3.3.6
mixer==7.2.1
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
zipp==3.7.0
